django_app/requirements.txt
asgiref==3.10.0
Django==4.2.25
filelock==3.20.0
fsspec==2025.10.0
gunicorn==23.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
